言語コード、例文、例文の訳の受け取りの実装 #79
