+++ src/atom/SentenceDiffState.ts
